octofit-tracker/backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
pymongo==4.5.0
mongoengine==0.27.0
